requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.1.0
databricks-sdk>=0.20.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
